refactor: 상품과 관련된 클래스들을 product 패키지로 이동
