+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,8 +1,8 @@
-package kitchenpos.application;
+package kitchenpos.product.application;
 
 import java.util.List;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
