+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
 import java.math.BigDecimal;
 import javax.persistence.Embedded;
@@ -6,6 +6,8 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import kitchenpos.domain.Name;
+import kitchenpos.domain.Price;
 
 @Entity
 public class Product {
