+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -11,15 +11,15 @@ import java.util.Optional;
 import kitchenpos.menu.fixture.MenuFixture;
 import kitchenpos.menu.fixture.MenuGroupFixture;
 import kitchenpos.menu.fixture.MenuProductFixture;
-import kitchenpos.application.fixture.ProductFixture;
+import kitchenpos.product.fixture.ProductFixture;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuProductRepository;
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.domain.Quantity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
